# swagger.yaml
openapi: 3.0.0
info:
  title: HRMS API
  version: 1.0.0
  description: API documentation for the HRMS (Human Resource Management System)

servers:
  - url: http://localhost:3000/api
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        userid:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        deptid:
          type: integer
        typeid:
          type: integer
      required:
        - firstname
        - lastname
        - email

    Department:
      type: object
      properties:
        deptid:
          type: integer
        deptname:
          type: string
        parentid:
          type: integer
      required:
        - deptname

paths:
  /users:
    get:
      summary: Get all users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalItems:
                        type: integer
                      totalPages:
                        type: integer

    post:
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully

  /departments:
    get:
      summary: Get all departments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'

    post:
      summary: Create a new department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        '201':
          description: Department created successfully